+++ lib/hashie.rb
@@ -21,15 +21,26 @@ module Hashie
     autoload :StringifyKeys,     'hashie/extensions/stringify_keys'
     autoload :SymbolizeKeys,     'hashie/extensions/symbolize_keys'
     autoload :DeepFetch,         'hashie/extensions/deep_fetch'
+    autoload :DeepFind,          'hashie/extensions/deep_find'
     autoload :PrettyInspect,     'hashie/extensions/pretty_inspect'
     autoload :KeyConversion,     'hashie/extensions/key_conversion'
+    autoload :MethodAccessWithOverride, 'hashie/extensions/method_access'
 
-    module Mash
-      autoload :ActiveModel,     'hashie/extensions/mash/active_model'
+    module Parsers
+      autoload :YamlErbParser, 'hashie/extensions/parsers/yaml_erb_parser'
     end
 
     module Dash
       autoload :IndifferentAccess, 'hashie/extensions/dash/indifferent_access'
     end
+
+    module Mash
+      autoload :SafeAssignment, 'hashie/extensions/mash/safe_assignment'
+    end
+  end
+
+  class << self
+    include Hashie::Extensions::StringifyKeys::ClassMethods
+    include Hashie::Extensions::SymbolizeKeys::ClassMethods
   end
 end