+++ lib/hashie/trash.rb
@@ -28,17 +28,31 @@ module Hashie
           fail ArgumentError, "Property name (#{property_name}) and :from option must not be the same"
         end
 
-        translations[options[:from]] = property_name
+        translations_hash[options[:from]] ||= {}
+        translations_hash[options[:from]][property_name] = options[:with] || options[:transform_with]
 
         define_method "#{options[:from]}=" do |val|
-          with = options[:with] || options[:transform_with]
-          self[property_name] = with.respond_to?(:call) ? with.call(val) : val
+          self.class.translations_hash[options[:from]].each do |name, with|
+            self[name] = with.respond_to?(:call) ? with.call(val) : val
+          end
         end
       else
         if options[:transform_with].respond_to? :call
           transforms[property_name] = options[:transform_with]
         end
       end
+    end
+
+    class << self
+      attr_reader :transforms, :translations_hash
+    end
+    instance_variable_set('@transforms', {})
+    instance_variable_set('@translations_hash', {})
+
+    def self.inherited(klass)
+      super
+      klass.instance_variable_set('@transforms', transforms.dup)
+      klass.instance_variable_set('@translations_hash', translations_hash.dup)
     end
 
     # Set a value on the Dash in a Hash-like way. Only works
@@ -58,7 +72,7 @@ module Hashie
     end
 
     def self.translation_exists?(name)
-      translations.key? name
+      translations_hash.key? name
     end
 
     def self.transformation_exists?(name)
@@ -71,20 +85,26 @@ module Hashie
 
     private
 
-    def self.properties
-      @properties ||= []
-    end
-
     def self.translations
-      @translations ||= {}
+      @translations ||= begin
+        h = {}
+        translations_hash.each do |(property_name, property_translations)|
+          h[property_name] = property_translations.size > 1 ? property_translations.keys : property_translations.keys.first
+        end
+        h
+      end
     end
 
     def self.inverse_translations
-      @inverse_translations ||= Hash[translations.map(&:reverse)]
-    end
-
-    def self.transforms
-      @transforms ||= {}
+      @inverse_translations ||= begin
+        h = {}
+        translations_hash.each do |(property_name, property_translations)|
+          property_translations.keys.each do |k|
+            h[k] = property_name
+          end
+        end
+        h
+      end
     end
 
     # Raises an NoMethodError if the property doesn't exist
@@ -100,7 +120,7 @@ module Hashie
     def initialize_attributes(attributes)
       return unless attributes
       attributes_copy = attributes.dup.delete_if do |k, v|
-        if self.class.translations.include?(k)
+        if self.class.translations_hash.include?(k)
           self[k] = v
           true
         end